Support notebook v4
The new ipython notebook version (v4) removes the `worksheets` structure, which
means that parsing a Bibliography notebook would fail.  This adds support for
v4, while maintaining support for v3 by checking for the existence of the older
structure.

Also allows for empty cells in the Bibliography notebook by checking if the
`cell_text` is defined before using it.

And fixes an apparent typo in the regex for removing `-->` from the cell_text.